vga: register size of text mode iomem
correctly, fix for "VGA: switch to configuration at runtime" commit.

The model uses the very same iomem for text mode as for graphical
mode, but unfortunately the whole size for the text mode was registered.
This lead to overlapping with the actual configured RAM, so that data written
by models (disk) was written to the vga iomem and the actual guest RAM was
empty ...