ps2keyboard: set to disabled by default
Fixes non-working PS2 emulation with linux guest, when used before in grub2
bootloader. GRUB2 uses the vbios_keyboard and consumes directly the host
input, and the ps2keyboard also consumes the host input, but it gets never
read. Noone seems to reset the ps2 keyboard to the default state after the
bootloader gives control to Linux.